Add ServerTiming parsing
